[MOL-20717][RP] move internal constants to separate data file
