+++ src/language-switcher/language-switcher.tsx
@@ -1,11 +1,8 @@
 import { announce } from "@react-aria/live-announcer";
 import { DropdownVariant } from "./dropdown-variant";
 import { LinkContainerVariant } from "./link-container-variant";
-import {
-    LANGUAGE_DISPLAY_MAP,
-    LanguageCode,
-    LanguageSwitcherProps,
-} from "./types";
+import { LANGUAGE_DISPLAY_MAP } from "./data";
+import { LanguageCode, LanguageSwitcherProps } from "./types";
 
 export const LanguageSwitcher = ({
     variant = "dropdown",
